requirements.txt
django
djangorestframework
drf-yasg
django-filter
uwsgi
requests
python-dotenv
boto3
